% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_defs}
\alias{ag_col_defs}
\title{Modify columns use vectorized.}
\usage{
ag_col_defs(x, columns = NULL, ...)
}
\arguments{
\item{x}{aggrid Object}

\item{columns}{Specify the column to modify, Default is all columns.}

\item{...}{please see \url{https://www.ag-grid.com/javascript-data-grid/column-properties/}.}
}
\description{
It is best to directly view the original document, please see \url{https://www.ag-grid.com/javascript-data-grid/column-properties/}.
}
\seealso{
\code{\link[=ag_col_def]{ag_col_def()}}
}
