% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_filter}
\alias{ag_col_filter}
\title{Columns: Filter}
\usage{
ag_col_filter(
  x,
  columns = NULL,
  filter = NULL,
  filterParams = NULL,
  filterValueGetter = NULL,
  getQuickFilterText = NULL,
  floatingFilter = NULL,
  floatingFilterComponent = NULL,
  floatingFilterComponentParams = NULL,
  suppressFiltersToolPanel = NULL
)
}
\arguments{
\item{x}{aggrid Object}

\item{columns}{Specify the column to modify, Default is all columns.}

\item{filter}{Filter component to use for this column. Set to true to use the default filter. Set to the name of a Provided Filter or set to a IFilterComp.See: Column Filters}

\item{filterParams}{Params to be passed to the filter component specified in filter.}

\item{filterValueGetter}{Function or expression. Gets the value for filtering purposes.More details}

\item{getQuickFilterText}{A function to tell the grid what Quick Filter text to use for this column if you don't want to use the default (which is calling toString on the value). More details}

\item{floatingFilter}{Whether to display a floating filter for this column.}

\item{floatingFilterComponent}{The custom component to be used for rendering the floating filter.  If none is specified the default AG Grid is used.}

\item{floatingFilterComponentParams}{Params to be passed to floatingFilterComponent.}

\item{suppressFiltersToolPanel}{Set to true if you do not want this column (filter) or group (filter group) to appear in the Filters Tool Panel.}
}
\description{
Columns: Filter
}
