% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_group}
\alias{ag_col_group}
\title{add aggrid header}
\usage{
ag_col_group(
  x,
  columns,
  headerName,
  groupId = NULL,
  marryChildren = NULL,
  suppressStickyLabel = NULL,
  openByDefault = NULL,
  columnGroupShow = NULL,
  toolPanelClass = NULL,
  suppressColumnsToolPanel = NULL,
  suppressFiltersToolPanel = NULL,
  suppressSpanHeaderHeight = NULL,
  tooltipComponent = NULL,
  tooltipComponentParams = NULL,
  headerClass = NULL,
  headerTooltip = NULL,
  headerGroupComponent = NULL,
  headerGroupComponentParams = NULL
)
}
\arguments{
\item{x}{aggrid Object}

\item{columns}{Specify the column to modify, Default is all columns.}

\item{columnGroupShow}{Whether to only show the column when the group is open / closed. If not set the column is always displayed as part of the group.}

\item{toolPanelClass}{CSS class to use for the tool panel cell. Can be a string, array of strings, or function.}

\item{suppressColumnsToolPanel}{Set to true if you do not want this column or group to appear in the Columns Tool Panel.}
}
\description{
add aggrid header
}
