% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_header}
\alias{ag_col_header}
\title{Columns: Header}
\usage{
ag_col_header(
  x,
  columns = NULL,
  headerName = NULL,
  headerValueGetter = NULL,
  headerTooltip = NULL,
  headerClass = NULL,
  headerComponent = NULL,
  headerComponentParams = NULL,
  wrapHeaderText = NULL,
  autoHeaderHeight = NULL,
  menuTabs = NULL,
  columnsMenuParams = NULL,
  suppressMenu = NULL,
  suppressHeaderKeyboardEvent = NULL,
  headerCheckboxSelection = NULL,
  headerCheckboxSelectionFilteredOnly = NULL,
  headerCheckboxSelectionCurrentPageOnly = NULL
)
}
\arguments{
\item{x}{aggrid Object}

\item{columns}{Specify the column to modify, Default is all columns.}

\item{headerName}{The name to render in the column header. If not specified and field is specified, the field name will be used as the header name.}

\item{headerValueGetter}{Function or expression. Gets the value for display in the header.More details}

\item{headerTooltip}{Tooltip for the column header}

\item{headerClass}{CSS class to use for the header cell. Can be a string, array of strings, or function. More details}

\item{headerComponent}{The custom header group component to be used for rendering the component header. If none specified the default AG Grid is used.See: Header Component}

\item{headerComponentParams}{The parameters to be passed to the headerComponent.}

\item{wrapHeaderText}{If enabled then column header names that are too long for the column width will wrap onto the next line. Default false}

\item{autoHeaderHeight}{If enabled then the column header row will automatically adjust height to accommodate the size of the header cell.  This can be useful when using your own headerComponent or long header names in conjunction with wrapHeaderText.}

\item{menuTabs}{Set to an array containing zero, one or many of the following options: 'filterMenuTab' | 'generalMenuTab' | 'columnsMenuTab'.  This is used to figure out which menu tabs are present and in which order the tabs are shown.}

\item{columnsMenuParams}{Params used to change the behaviour and appearance of the Columns Menu tab. More details}

\item{suppressMenu}{Set to true if no menu should be shown for this column header.}

\item{suppressHeaderKeyboardEvent}{Suppress the grid taking action for the relevant keyboard event when a header is focused. More details}

\item{headerCheckboxSelection}{If true or the callback returns true, a 'select all' checkbox will be put into the header. More details}

\item{headerCheckboxSelectionFilteredOnly}{If true, the header checkbox selection will only select filtered items.}

\item{headerCheckboxSelectionCurrentPageOnly}{If true, the header checkbox selection will only select nodes on the current page.}
}
\description{
Columns: Header
}
