% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_pinned}
\alias{ag_col_pinned}
\title{Columns: Pinned}
\usage{
ag_col_pinned(
  x,
  columns = NULL,
  pinned = NULL,
  initialPinned = NULL,
  lockPinned = NULL
)
}
\arguments{
\item{x}{aggrid Object}

\item{columns}{Specify the column to modify, Default is all columns.}

\item{pinned}{Pin a column to one side: right or left.}

\item{initialPinned}{Same as pinned, except only applied when creating a new column. Not applied when updating column definitions.}

\item{lockPinned}{Set to true to block the user pinning the column, the column can only be pinned via definitions or API.}
}
\description{
Columns: Pinned
}
