% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_render}
\alias{ag_col_render}
\title{Columns cell Render}
\usage{
ag_col_render(
  x,
  columns = NULL,
  cellStyle = NULL,
  cellClass = NULL,
  cellClassRules = NULL,
  cellRenderer = NULL,
  cellRendererParams = NULL,
  cellRendererSelector = NULL,
  autoHeight = NULL,
  wrapText = NULL,
  enableCellChangeFlash = NULL,
  suppressCellFlash = NULL
)
}
\arguments{
\item{cellStyle}{An object of css values / or function returning an object of css values for a particular cell. More details}

\item{cellClass}{Class to use for the cell. Can be string, array of strings, or function that returns a string or array of strings. More details}

\item{cellClassRules}{Rules which can be applied to include certain CSS classes. More details}

\item{cellRenderer}{Provide your own cell Renderer component for this column's cells.See: Cell Renderer}

\item{cellRendererParams}{Params to be passed to the cellRenderer component.}

\item{cellRendererSelector}{Callback to select which cell renderer to be used for a given row within the same column. More details}

\item{autoHeight}{Set to true to have the grid calculate the height of a row based on contents of this column.}

\item{wrapText}{Set to true to have the text wrap inside the cell - typically used with autoHeight.}

\item{enableCellChangeFlash}{Set to true to flash a cell when it's refreshed.}

\item{suppressCellFlash}{Set to true to prevent this column from flashing on changes. Only applicable if cell flashing is turned on for the grid.}
}
\description{
Columns cell Render
}
