% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_sort}
\alias{ag_col_sort}
\title{Custom columns sort}
\usage{
ag_col_sort(
  x,
  columns = NULL,
  sortable = NULL,
  sort = NULL,
  initialSort = NULL,
  sortIndex = NULL,
  initialSortIndex = NULL,
  sortingOrder = NULL,
  comparator = NULL,
  unSortIcon = NULL,
  ...
)
}
\arguments{
\item{x}{aggrid Object}

\item{columns}{Specify the column to modify, Default is all columns.}

\item{sortable}{Set to true to allow sorting on this column.}

\item{sort}{If sorting by default, set it here. Set to asc or desc. More details}

\item{initialSort}{Same as sort, except only applied when creating a new column. Not applied when updating column definitions. More details}

\item{sortIndex}{If sorting more than one column by default, specifies order in which the sorting should be applied.}

\item{initialSortIndex}{Same as sortIndex, except only applied when creating a new column. Not applied when updating column definitions.}

\item{sortingOrder}{Array defining the order in which sorting occurs (if sorting is enabled). An array with any of the following in any order \link{'asc','desc',null}.}

\item{comparator}{Override the default sorting order by providing a custom sort comparator.  valueA, valueB are the values to compare.  nodeA,  nodeB are the corresponding RowNodes. Useful if additional details are required by the sort.  isDescending - true if sort direction is desc. Not to be used for inverting the return value as the grid already applies asc or desc ordering.}

\item{unSortIcon}{Set to true if you want the unsorted icon to be shown when no sort is applied to this column.}

\item{...}{please see \url{https://www.ag-grid.com/javascript-data-grid/column-properties/}.}
}
\description{
Custom columns sort
}
