% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_span}
\alias{ag_col_span}
\title{Columns: Spanning}
\usage{
ag_col_span(x, columns = NULL, colSpan = NULL, rowSpan = NULL, ...)
}
\arguments{
\item{x}{aggrid Object}

\item{columns}{Specify the column to modify, Default is all columns.}

\item{colSpan}{By default, each cell will take up the width of one column. You can change this behaviour to allow cells to span multiple columns.}

\item{rowSpan}{By default, each cell will take up the height of one row. You can change this behaviour to allow cells to span multiple rows.}

\item{...}{please see \url{https://www.ag-grid.com/javascript-data-grid/column-properties/}.}
}
\description{
Columns: Spanning
}
