% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_tooltip}
\alias{ag_col_tooltip}
\title{Custom columns cell tooltip}
\usage{
ag_col_tooltip(
  x,
  columns = NULL,
  tooltipField = NULL,
  tooltipValueGetter = NULL,
  tooltipComponent = NULL,
  tooltipComponentParams = NULL,
  ...
)
}
\arguments{
\item{x}{aggrid Object}

\item{columns}{Specify the column to modify, Default is all columns.}

\item{tooltipField}{The field of the tooltip to apply to the cell.}

\item{tooltipValueGetter}{Callback that should return the string to use for a tooltip, \code{tooltipField} takes precedence if set. If using a custom \code{tooltipComponent} you may return any custom value to be passed to your tooltip component.}

\item{tooltipComponent}{Provide your own tooltip component for the column.}

\item{...}{please see \url{https://www.ag-grid.com/javascript-data-grid/column-properties/}.}

\item{The}{params used to configure \code{tooltipComponent}.}
}
\description{
Custom columns cell tooltip
}
