% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_col_width}
\alias{ag_col_width}
\title{Custom colums width}
\usage{
ag_col_width(
  x,
  columns = NULL,
  width = NULL,
  initialWidth = NULL,
  minWidth = NULL,
  maxWidth = NULL,
  flex = NULL,
  initialFlex = NULL,
  resizable = NULL,
  suppressSizeToFit = NULL,
  suppressAutoSize = NULL
)
}
\arguments{
\item{x}{aggrid Object}

\item{columns}{Specify the column to modify, Default is all columns.}

\item{width}{Initial width in pixels for the cell.}

\item{initialWidth}{Same as width, except only applied when creating a new column. Not applied when updating column definitions.}

\item{minWidth}{Minimum width in pixels for the cell.}

\item{maxWidth}{Maximum width in pixels for the cell.}

\item{flex}{Used instead of width when the goal is to fill the remaining empty space of the grid.}

\item{initialFlex}{Same as flex, except only applied when creating a new column. Not applied when updating column definitions.}

\item{resizable}{Set to true to allow this column should be resized.}

\item{suppressSizeToFit}{Set to true if you want this column's width to be fixed during 'size to fit' operations.}

\item{suppressAutoSize}{Set to true if you do not want this column to be auto-resizable by double clicking it's edge.}
}
\description{
More message: \url{https://www.ag-grid.com/javascript-data-grid/column-properties/#reference-width}
}
\seealso{
\code{\link[=ag_col_def]{ag_col_def()}}
}
