% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_size_columns}
\alias{ag_size_columns}
\title{Resize all column by columns content}
\usage{
ag_size_columns(x, skipHeader = FALSE)
}
\arguments{
\item{x}{aggrid Object}

\item{skipHeader}{Enable skip Header?}
}
\description{
Resize all column by columns content
}
\examples{

aggrid(iris) |>
  ag_size_columns()
}
