% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{ag_size_fit}
\alias{ag_size_fit}
\title{Resize all column by container width}
\usage{
ag_size_fit(x)
}
\arguments{
\item{x}{aggrid Object}
}
\description{
Example from \url{https://www.ag-grid.com/javascript-data-grid/column-sizing/#resizing-example}
}
