% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrid-package.R
\docType{package}
\name{aggrid-package}
\alias{aggrid}
\alias{aggrid-package}
\title{aggrid: Html widget interface the ag-grid library}
\description{
Html widgets table.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://liripo.github.io/aggrid/}
}

}
\author{
\strong{Maintainer}: Liripo (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
