% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_utils.R
\name{auto_size_columns}
\alias{auto_size_columns}
\title{Resize all column by columns content}
\usage{
auto_size_columns(x, skipHeader = FALSE)
}
\arguments{
\item{x}{aggrid Object}

\item{skipHeader}{Enable skip Header?}
}
\description{
Resize all column by columns content
}
