% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_aggrid_filter}
\alias{to_aggrid_filter}
\title{R column class to aggrid filter}
\source{
\url{https://www.ag-grid.com/javascript-data-grid/filtering-overview/}
}
\usage{
to_aggrid_filter(type)
}
\description{
R column class to aggrid filter
}
