% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrid-package.R
\docType{package}
\name{aggrid-package}
\alias{aggrid}
\alias{aggrid-package}
\title{aggrid: Html widget interface the ag-grid library}
\description{
Html widgets table.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://liripo.github.io/aggrid/,https://github.com/Liripo/aggrid}
  \item Report bugs at \url{https://github.com/Liripo/aggrid/issues}
}

}
\author{
\strong{Maintainer}:  Liripo \email{liripo@qq.com} (\href{https://orcid.org/0009-0007-6069-1375}{ORCID})

}
\keyword{internal}
