% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrid-package.R
\docType{package}
\name{aggrid-package}
\alias{aggrid}
\alias{aggrid-package}
\title{aggrid: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://liripo.github.io/aggrid/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
