% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrid.R
\name{aggrid}
\alias{aggrid}
\title{Create a HTML widget using the ag-grid library}
\usage{
aggrid(
  data,
  rowSelection = c("multiple", "single"),
  suppressRowClickSelection = FALSE,
  selectedRows = NULL,
  checkboxSelection = FALSE,
  pagination = FALSE,
  paginationPageSize = 10,
  ...,
  theme = "ag-theme-balham",
  community = FALSE,
  use_cdn = FALSE,
  server = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{data frame or matrix.}

\item{rowSelection}{three options：\code{multiple},\code{single}.}

\item{checkboxSelection}{Enable checkbox in first column?}

\item{pagination}{Enable pagination allows?}

\item{paginationPageSize}{default size is \code{10}.}

\item{...}{For more options, please see \url{https://www.ag-grid.com/javascript-data-grid/grid-options}}

\item{theme}{Specify theme. Default is \code{ag-theme-balham}. please see \url{https://www.ag-grid.com/example/?theme=ag-theme-alpine}}

\item{community}{Enable to use community?}

\item{use_cdn}{Depends resource to use cdn?}

\item{server}{server model}

\item{width, height}{Width/Height in pixels (optional, defaults to automatic sizing)}

\item{elementId}{An id for the widget (a random string by default).}
}
\description{
This function creates a HTML widget to display matrix or a data frame using ag-grid.
}
\examples{
aggrid(iris)

## use theme
aggrid(iris, theme = "ag-theme-alpine")

## auto size columns
aggrid(iris) |> auto_size_columns()
}
